# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Intell Meet API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new meeting room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks for Kanban board
      responses:
        "200":
          description: Tasks retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new Kanban task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a Kanban task (status, description, assignee, etc.)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a Kanban task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /participants/mute:
    post:
      operationId: muteParticipant
      tags: [meetings]
      summary: Toggle mute status of a participant (Host only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MuteParticipantInput"
      responses:
        "200":
          description: Participant mute updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
  /participants/remove:
    post:
      operationId: removeParticipant
      tags: [meetings]
      summary: Remove a participant from the meeting (Host only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveParticipantInput"
      responses:
        "200":
          description: Participant removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]
  /participants/raise-hand:
    post:
      operationId: raiseHandParticipant
      tags: [meetings]
      summary: Toggle hand-raised status of a participant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaiseHandInput"
      responses:
        "200":
          description: Hand-raised status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Admin, Manager, Member]
        phoneNumber:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        authProvider:
          type: string
          enum: [local, google]
        googleId:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        hasPassword:
          type: boolean
        notificationSettings:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
            sms:
              type: boolean
          required: [email, push, sms]
        createdAt:
          type: string
      required: [id, name, email, role, createdAt, authProvider, emailVerified, hasPassword]
...
    Room:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
        participantCount:
          type: integer
      required: [id, name, createdAt, participantCount]
    RoomInput:
      type: object
      properties:
        name:
          type: string
      required: [name]
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Todo, In Progress, Done]
        assignee:
          $ref: "#/components/schemas/User"
        dueDate:
          type: ["string", "null"]
        teamId:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, description, status, dueDate, teamId, createdAt]
    TaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Todo, In Progress, Done]
        assigneeId:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        teamId:
          type: ["string", "null"]
      required: [title]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Todo, In Progress, Done]
        assigneeId:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
...
    MuteParticipantInput:
      type: object
      properties:
        meetingId:
          type: string
        userId:
          type: string
        isMuted:
          type: boolean
      required: [meetingId, userId, isMuted]
    RemoveParticipantInput:
      type: object
      properties:
        meetingId:
          type: string
        userId:
          type: string
      required: [meetingId, userId]
    RaiseHandInput:
      type: object
      properties:
        meetingId:
          type: string
        userId:
          type: string
        isRaisedHand:
          type: boolean
      required: [meetingId, userId, isRaisedHand]
...
    Participant:
      type: object
      properties:
        id:
          type: string
        meetingId:
          type: string
        userId:
          type: ["string", "null"]
        displayName:
          type: string
        role:
          type: string
          enum: [host, co-host, participant]
        status:
          type: string
          enum: [waiting, admitted, rejected, left]
        isMuted:
          type: boolean
        isCameraOff:
          type: boolean
        isRaisedHand:
          type: boolean
        joinedAt:
          type: string
        leftAt:
          type: ["string", "null"]
      required: [id, meetingId, displayName, role, status, isMuted, isCameraOff, isRaisedHand, joinedAt]
...